♻️ Refactor: like/dislike 버튼 utils 분리
